+++ arrus/core/devices/us4r/probeadapter/ProbeAdapterImplTest.cpp
@@ -105,6 +105,10 @@ public:
     MOCK_METHOD(Ius4OEMRawHandle, getIUs4oem, (), (override));
     MOCK_METHOD(void, enableSequencer, (), (override));
     MOCK_METHOD(std::vector<uint8_t>, getChannelMapping, (), (override));
+    MOCK_METHOD(void, checkFirmwareVersion, (), (override));
+    MOCK_METHOD(void, checkState, (), (override));
+    MOCK_METHOD(uint32, getFirmwareVersion, (), (override));
+    MOCK_METHOD(uint32, getTxFirmwareVersion, (), (override));
 };
 
 class AbstractProbeAdapterImplTest : public ::testing::Test {